Create README.md
Background: 
Our client in the fintech space was on a mission to help small and microlenders make informed credit decisions. They approached us with a request to create a data visualization dashboard that they could embed in their portal.

Problem: 
The client had already invested in a tool that lacked the necessary visualization capabilities and interactive features to meet their needs. The learning curve to get familiar with the tool was steep and some crucial features were missing.

Solution: 
Despite these challenges, we put our expertise in dashboard development to the test and delivered six compelling charts within a week. Our team leveraged years of knowledge in the field to create a visually appealing and informative solution.

Result: 
The client was thrilled with our work and quickly moved to conduct a beta test with their long-term clients. The feedback was overwhelmingly positive, with the only comment being that the dashboard couldn't have been any better. The client was so impressed that they immediately integrated our solution into their product without any alterations. We were grateful for the opportunity to partner with our client and look forward to supporting them with future requirements.